jobs-fetching-service/requirements.txt
beautifulsoup4==4.12.3
soupsieve==2.6
pyyaml==6.0.2